requirements.txt
allure-pytest==2.14.1
allure-python-commons==2.14.1
attrs==25.4.0
colorama==0.4.6
execnet==2.1.2
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pytest==8.3.3
pytest-xdist==3.6.1
